Screen selection for capturing now works with multi-head
i.e. multiple monitors
First, tries to find ‘Entire screen’, which is the name of the only monitor in Mac setups, then tries to take ‘Screen 1’, the first monitor in multi-head setups on mac.
If that also does not find a suitable screen, it selects the first source obtained from `desktopCapturer.getSources`